Separate codec into codec crate
