+++ flutter-engine-codec/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "flutter-engine-codec"
+description = "Flutter engine codec interop"
+version = "0.3.5"
+authors = ["juju <gliheng@gmail.com>", "Sophie Tauchert <sophie@999eagle.moe>", "Chandler Newman <chandler2newman@gmail.com>"]
+keywords = ["flutter"]
+license = "MIT"
+repository = "https://github.com/gliheng/flutter-rs"
+edition = "2018"
+
+[dependencies]
+flutter-engine-sys = { version = "0.3.5", path = "../flutter-engine-sys" }
+serde = { version = "^1.0", features = ["derive"] }
+serde_json = "^1.0"
+log = "^0.4"
