+++ flutter-engine-codec/src/macros.rs
@@ -3,7 +3,7 @@ macro_rules! json_value {
     ($($tt:tt)*) => {
         {
             use std::convert::TryInto;
-            serde_json::json!($($tt)*).try_into().unwrap_or($crate::codec::Value::Null)
+            serde_json::json!($($tt)*).try_into().unwrap_or($crate::Value::Null)
         }
     };
 }
