+++ flutter-engine-codec/src/string_codec.rs
