+++ flutter-engine/Cargo.toml
@@ -11,6 +11,7 @@ edition = "2018"
 
 [dependencies]
 flutter-engine-sys = { version = "0.3.5", path = "../flutter-engine-sys" }
+flutter-engine-codec = { version = "0.3.5", path = "../flutter-engine-codec" }
 glfw = "^0.28"
 gl = "0.11.0"
 libc = "0.2.44"
