+++ flutter-engine/src/error.rs
@@ -3,7 +3,7 @@ use std::{
     sync::mpsc::{RecvError, SendError},
 };
 
-use crate::codec::{MethodCallResult, Value};
+use flutter_engine_codec::{MethodCallResult, Value, error::ValueError};
 
 #[derive(Debug)]
 pub enum MethodArgsError {
@@ -198,30 +198,3 @@ impl Into<MethodCallResult> for MethodCallError {
         }
     }
 }
-
-#[derive(Debug)]
-pub enum ValueError {
-    Message(String),
-    WrongType,
-    NoList,
-    NoMap,
-}
-
-impl fmt::Display for ValueError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            ValueError::Message(s) => write!(f, "{}", s),
-            ValueError::WrongType => write!(f, "wrong type"),
-            ValueError::NoList => write!(f, "value is not a list"),
-            ValueError::NoMap => write!(f, "value is not a map"),
-        }
-    }
-}
-
-impl serde::de::Error for ValueError {
-    fn custom<T: fmt::Display>(msg: T) -> Self {
-        ValueError::Message(msg.to_string())
-    }
-}
-
-impl error::Error for ValueError {}
