+++ flutter-engine/src/plugins/prelude.rs
@@ -5,18 +5,18 @@ pub use std::{
 
 pub use serde::{Deserialize, Serialize};
 
+pub use flutter_engine_codec::{
+    json_codec, json_value, standard_codec, string_codec, value::from_value, MethodCall,
+    MethodCallResult, Value,
+};
+
 pub use crate::{
     channel::{
         BasicMessageChannel, ChannelRegistrar, EventChannel, EventHandler, JsonMethodChannel,
         MessageChannel, MessageHandler, MethodCallHandler, MethodChannel, StandardMethodChannel,
     },
-    codec::{
-        json_codec, standard_codec, string_codec, value::from_value, MethodCall, MethodCallResult,
-        Value,
-    },
     error::{MessageError, MethodArgsError, MethodCallError},
     ffi::PlatformMessageResponseHandle,
-    json_value,
     plugins::{Plugin, PluginRegistrar},
     RuntimeData, Window,
 };
